+++ CPP-Module-02/ex02/Fixed.cpp
@@ -6,7 +6,7 @@
 /*   By: corozco <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/23 14:15:38 by corozco           #+#    #+#             */
-/*   Updated: 2021/01/24 12:22:27 by corozco          ###   ########.fr       */
+/*   Updated: 2021/02/28 15:01:21 by corozco          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ Fixed::Fixed( const int tmp )
 
 Fixed::Fixed( const float tmp)
 {
-	this->m_fixed = round(tmp * (1 << this->m_bits));
+	this->m_fixed = roundf(tmp * (1 << this->m_bits));
 }
 
 Fixed::Fixed( Fixed const & obj )
@@ -61,7 +61,7 @@ float	Fixed::toFloat( void ) const
 
 int		Fixed::toInt( void ) const
 {
-	return ( (int)round(this->m_fixed >> this->m_bits) );
+	return ( (int)roundf(this->m_fixed >> this->m_bits) );
 }
 
 std::ostream &	operator<<( std::ostream & o, Fixed const & i )
